src/requirements.txt
aiohttp==3.11.2
config==0.5.1
database==0.6.0
easyocr==1.7.2
keyboard==0.13.5
mysql_replication==1.0.9
numpy==2.2.2
opencv_python==4.8.0.76
opencv_python_headless==4.10.0.84
Pillow==10.0.0
Pillow==11.1.0
PyAutoGUI==0.9.54
PyQt5==5.15.11
PyQt5_sip==12.15.0
python-dotenv==1.0.1
pywinauto==0.6.8
Requests==2.32.3
SQLAlchemy==2.0.36
